requirements.txt
psutil==5.8.0
pytest==6.2.4
PyYAML==5.4.1
requests==2.25.1
tabulate==0.8.9
